project/api/serializers/ingredient_serializer.py
from rest_framework import serializers

from recipes.models import Ingredient


class IngredientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Ingredient
        fields = ['title', 'dimension']
        read_only_fields = ['title', 'dimension']
