requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
gunicorn==20.1.0
isort==5.9.2
mccabe==0.6.1
pep8-naming==0.12.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.1
pytils==0.3
pytz==2021.1
reportlab==3.5.68
sorl-thumbnail==12.7.0
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
